src/python/requirements.txt
numpy
tensorflow
keras.datasets
scipy.ndimage